Covid Management System
The project “Covid Management System” is to be developed for maintaining the Covid details like patient details,lab technician details,testing details,live updates etc.The system is efficient in generating reports which will help in the Covid Management up-to-date easily. The objective of this project is to develop a system that automates the processes and activities of Covid. The purpose is to design a system using which one can perform all operations related to Covid testing and vaccination.